wip: combine into default press button type
