+++ lib/Controls/ControlTypes/Button/Press.js
@@ -3,7 +3,8 @@ import Registry from '../../../Registry.js'
 import { cloneDeep } from 'lodash-es'
 
 /**
- * Class for the press button control.
+ * Class for the default button control.
+ * This button supports the default press behaviour, and timed releases
  *
  * @extends ButtonControlBase
  * @author Håkon Nessjøen <haakon@bitfocus.io>
@@ -24,7 +25,11 @@ import { cloneDeep } from 'lodash-es'
  * disclosing the source code of your own applications.
  */
 export default class ControlButtonPress extends ButtonControlBase {
-	type = 'press'
+	/**
+	 * Last down press from a deviceId
+	 * @access private
+	 */
+	#lastDownForDevices = {}
 
 	/**
 	 * @param {Registry} registry - the application core
@@ -35,8 +40,11 @@ export default class ControlButtonPress extends ButtonControlBase {
 	constructor(registry, controlId, storage, isImport) {
 		super(registry, controlId, 'press-button', 'Controls/Button/Press')
 
+		this.type = 'press'
+
 		this.style = cloneDeep(ButtonControlBase.DefaultStyle)
 		this.options = cloneDeep(ButtonControlBase.DefaultOptions)
+
 		this.feedbacks.feedbacks = []
 		this.actions.action_sets = {
 			down: [],
@@ -59,6 +67,95 @@ export default class ControlButtonPress extends ButtonControlBase {
 
 			if (isImport) this.postProcessImport()
 		}
+	}
+
+	/**
+	 * Add an action set to this control
+	 * @returns {boolean} success
+	 * @access public
+	 */
+	actionSetAdd() {
+		let redraw = false
+
+		const existingKeys = Object.keys(this.actions.action_sets)
+			.filter((k) => !isNaN(k))
+			.map((k) => Number(k))
+		if (existingKeys.length === 0) {
+			// add the default '1000' set
+			this.actions.action_sets['1000'] = []
+			redraw = true
+		} else {
+			// add one after the last
+			const max = Math.max(...existingKeys)
+			const newIndex = Math.floor(max / 1000) * 1000 + 1000
+			this.actions.action_sets[newIndex] = []
+		}
+
+		this.commitChange(redraw)
+
+		return true
+	}
+
+	/**
+	 * Remove an action-set from this control
+	 * @param {string} setId the id of the action-set
+	 * @returns {boolean} success
+	 * @access public
+	 */
+	actionSetRemove(setId) {
+		// Ensure valid
+		if (isNaN(setId)) return false
+
+		const oldKeys = Object.keys(this.actions.action_sets)
+
+		if (oldKeys.length > 1) {
+			let redraw = false
+
+			if (this.actions.action_sets[setId]) {
+				// Inform modules of the change
+				for (const action of this.actions.action_sets[setId]) {
+					this.actions.cleanupAction(action)
+				}
+
+				// Assume it exists
+				delete this.actions.action_sets[setId]
+
+				// Check if the status has changed
+				if (this.checkButtonStatus()) redraw = true
+
+				// Save the change, and perform a draw
+				this.commitChange(redraw)
+
+				return true
+			}
+		}
+
+		return false
+	}
+
+	/**
+	 * Rename an action-sets
+	 * @param {string} oldSetId The old id of the set
+	 * @param {string} newSetId The new id for the set
+	 * @returns {boolean} success
+	 * @access public
+	 */
+	actionSetRename(oldSetId, newSetId) {
+		// Only valid when both are numbers
+		if (isNaN(newSetId) || isNaN(oldSetId)) return false
+
+		// Ensure old set exists
+		if (!this.actions.action_sets[oldSetId]) return false
+
+		// Ensure new set doesnt already exist
+		if (this.actions.action_sets[newSetId]) return false
+
+		this.actions.action_sets[newSetId] = this.actions.action_sets[oldSetId]
+		delete this.actions.action_sets[oldSetId]
+
+		this.commitChange(false)
+
+		return true
 	}
 
 	/**
@@ -91,21 +188,46 @@ export default class ControlButtonPress extends ButtonControlBase {
 	 * @access public
 	 */
 	pressControl(pressed, deviceId) {
+		let pressedDuration = 0
+		if (deviceId) {
+			// Calculate the press duration, or track when the press started
+			if (pressed) {
+				this.#lastDownForDevices[deviceId] = Date.now()
+			} else {
+				const lastDown = this.#lastDownForDevices[deviceId]
+				delete this.#lastDownForDevices[deviceId]
+				if (lastDown) {
+					pressedDuration = Date.now() - lastDown
+				}
+			}
+		}
+
 		const changed = this.setPushed(pressed, deviceId)
 
 		// if the state has changed, the choose the set to execute
-		// if (changed) { // TODO - this would be good but 'breaks' old behaviour
-		const action_set_id = pressed ? 'down' : 'up'
+		if (changed) {
+			let action_set_id = pressed ? 'down' : 'up'
 
-		const actions = this.actions.action_sets[action_set_id]
-		if (actions) {
-			this.logger.silly('found actions')
+			if (!pressed && pressedDuration) {
+				// find the correct set to execute on up
 
-			this.controls.actions.runMultipleActions(actions, this.controlId, this.options.relativeDelay, {
-				deviceid: deviceId,
-			})
+				const setIds = Object.keys(this.actions.action_sets)
+					.map((id) => Number(id))
+					.filter((id) => !isNaN(id) && id < pressedDuration)
+				if (setIds.length) {
+					action_set_id = Math.max(...setIds)
+				}
+			}
+
+			const actions = this.actions.action_sets[action_set_id]
+			if (actions) {
+				this.logger.silly('found actions')
+
+				this.controls.actions.runMultipleActions(actions, this.controlId, this.options.relativeDelay, {
+					deviceid: deviceId,
+				})
+			}
 		}
-		// }
 	}
 
 	/**
