+++ lib/Controls/Controller.js
@@ -13,7 +13,6 @@ import ActionRecorder from './ActionRecorder.js'
 import ControlTrigger from './ControlTypes/Triggers/Trigger.js'
 import { nanoid } from 'nanoid'
 import TriggerEvents from './TriggerEvents.js'
-import ControlButtonTimed from './ControlTypes/Button/Timed.js'
 
 export const TriggersListRoom = 'triggers:list'
 
@@ -697,8 +696,6 @@ class ControlsController extends CoreBase {
 					return new ControlButtonPress(this.registry, controlId, controlObj, isImport)
 				case 'step':
 					return new ControlButtonStep(this.registry, controlId, controlObj, isImport)
-				case 'timed':
-					return new ControlButtonTimed(this.registry, controlId, controlObj, isImport)
 				case 'pagenum':
 					return new ControlButtonPageNumber(this.registry, controlId, controlObj, isImport)
 				case 'pageup':
